+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # master
 
 - Add `defaultSearch` prop
+- Add `preferredSize` prop, which always downloads photos at the requested size